[REFACTORING] Split MailboxSetMethod and extract create/update/destroy
